TeamsBot/src/requirements.txt
botbuilder-core==4.14.4
aiohttp==3.8.4